# Car-Wash-Api-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Gateway Booking Service",
        "title": "Api Gateway",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/providers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all providers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Get all providers",
                "parameters": [
                    {
                        "type": "number",
                        "description": "averageRating",
                        "name": "averageRating",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "companyName",
                        "name": "companyName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carwash.ListProvidersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Create a new provider",
                "parameters": [
                    {
                        "description": "provider",
                        "name": "provider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/carwash.CreateProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carwash.CreateProviderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/providers/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search providers by company name or description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Search providers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company Name",
                        "name": "CompanyName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Description",
                        "name": "Description",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carwash.SearchProvidersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/providers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Get a provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carwash.GetProviderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Update a provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "averageRating",
                        "name": "averageRating",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "companyName",
                        "name": "companyName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "serviceId",
                        "name": "serviceId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "availability",
                        "name": "availability",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carwash.UpdateProviderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Delete a provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carwash.DeleteProviderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "carwash.Availability": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "carwash.CreateProviderRequest": {
            "type": "object",
            "properties": {
                "availability": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/carwash.Availability"
                    }
                },
                "average_rating": {
                    "type": "number"
                },
                "company_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/carwash.GeoPoing"
                },
                "service_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "carwash.CreateProviderResponse": {
            "type": "object"
        },
...
        "carwash.DeleteProviderResponse": {
            "type": "object"
        },
...
        "carwash.GeoPoing": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "carwash.GetProviderResponse": {
            "type": "object",
            "properties": {
                "provider": {
                    "$ref": "#/definitions/carwash.Provider"
                }
            }
        },
...
        "carwash.ListProvidersResponse": {
            "type": "object",
            "properties": {
                "provider": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/carwash.Provider"
                    }
                }
            }
        },
...
        "carwash.Provider": {
            "type": "object",
            "properties": {
                "availability": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/carwash.Availability"
                    }
                },
                "average_rating": {
                    "type": "number"
                },
                "company_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/carwash.GeoPoing"
                },
                "service_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "carwash.SearchProvidersResponse": {
            "type": "object",
            "properties": {
                "providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/carwash.Provider"
                    }
                }
            }
        },
...
        "carwash.UpdateProviderResponse": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}